Initial seperation of of REST from UI
