+++ src/components/OSCALControlImplementation.js
@@ -64,7 +64,6 @@ export default function OSCALControlImplementation(props) {
                       isEditable={props.isEditable}
                       onFieldSave={props.onFieldSave}
                       partialRestData={props.partialRestData}
-                      restPath={props.restPath}
                     />
                   )
                 )}
