+++ src/components/OSCALLoader.js
@@ -4,7 +4,7 @@ import Split from "react-split";
 import { makeStyles } from "@material-ui/core/styles";
 import { Box, Fab } from "@material-ui/core";
 import CodeIcon from "@material-ui/icons/Code";
-import { populatePartialRestData } from "./oscal-utils/OSCALUtils";
+import { populatePartialRestData, buildRestRequestUrl, restMethods, oscalObjectTypes } from "./oscal-utils/OSCALRestUtils";
 import ErrorBoundary, { BasicError } from "./ErrorBoundary";
 import OSCALSsp from "./OSCALSsp";
 import OSCALCatalog from "./OSCALCatalog";
@@ -12,41 +12,6 @@ import OSCALComponentDefinition from "./OSCALComponentDefinition";
 import OSCALProfile from "./OSCALProfile";
 import OSCALLoaderForm from "./OSCALLoaderForm";
 import OSCALJsonEditor from "./OSCALJsonEditor";
-
-const oscalObjectTypes = {
-  catalog: {
-    name: "Catalog",
-    defaultUrl:
-      "https://raw.githubusercontent.com/EasyDynamics/oscal-demo-content/main/catalogs/NIST_SP-800-53_rev5_catalog.json",
-    defaultUuid: "613fca2d-704a-42e7-8e2b-b206fb92b456",
-    jsonRootName: "catalog",
-    restPath: "catalogs",
-  },
-  component: {
-    name: "Component",
-    defaultUrl:
-      "https://raw.githubusercontent.com/EasyDynamics/oscal-demo-content/main/component-definitions/example-component.json",
-    defaultUuid: "8223d65f-57a9-4689-8f06-2a975ae2ad72",
-    jsonRootName: "component-definition",
-    restPath: "component-definitions",
-  },
-  profile: {
-    name: "Profile",
-    defaultUrl:
-      "https://raw.githubusercontent.com/EasyDynamics/oscal-demo-content/main/profiles/NIST_SP-800-53_rev4_MODERATE-baseline_profile.json",
-    defaultUuid: "8b3beca1-fcdc-43e0-aebb-ffc0a080c486",
-    jsonRootName: "profile",
-    restPath: "profiles",
-  },
-  ssp: {
-    name: "SSP",
-    defaultUrl:
-      "https://raw.githubusercontent.com/EasyDynamics/oscal-demo-content/main/system-security-plans/ssp-example.json",
-    defaultUuid: "cff8385f-108e-40a5-8f7a-82f3dc0eaba8",
-    jsonRootName: "system-security-plan",
-    restPath: "system-security-plans",
-  },
-};
 
 const useStyles = makeStyles((theme) => ({
   split: {
@@ -120,37 +85,16 @@ export default function OSCALLoader(props) {
       );
   };
 
-  /**
-   * Sends a REST request of type restMethod to a backend service and updates the viewer if
-   * the request is successful.
-   *
-   * @param appendToLastFieldInPath boolean indicating if the updated value should be appended to an array or replace an existing value
-   * @param partialRestData data that will be passed into the body of the REST request, may not initially contain the updates
-   * @param editedFieldJsonPath path to the field that is being updated
-   * @param newValue updated value for the edited field
-   * @param httpMethod the HTTP request type
-   * @param restUrlPath path defining where in the file the modifications are made
-   * @param jsonRootName root OSCAL object, as it appears on the corresponding object file, of the JSON file
-   * @param restPath main url path for access the OSCAL files in REST mode
-   */
-  const handleRestRequest = (
+  const handleFieldSave = (
     appendToLastFieldInPath,
     partialRestData,
     editedFieldJsonPath,
     newValue,
-    httpMethod,
     restUrlPath,
-    jsonRootName = null,
-    restPath = null
+    oscalObjectType
   ) => {
-    let url;
-    if (!restUrlPath || restUrlPath === "") {
-      url = `${process.env.REACT_APP_REST_BASE_URL}/${restPath}/${partialRestData[jsonRootName].uuid}`;
-    } else if (restUrlPath.startsWith("http", 0)) {
-      url = restUrlPath;
-    } else {
-      url = `${process.env.REACT_APP_REST_BASE_URL}/${restUrlPath}`;
-    }
+    const requestUrl = buildRestRequestUrl(partialRestData, restUrlPath, oscalObjectType);
+
     if (newValue) {
       populatePartialRestData(
         partialRestData,
@@ -160,33 +104,23 @@ export default function OSCALLoader(props) {
       );
     }
 
-    setIsLoaded(false);
-    setIsResolutionComplete(false);
-
-    const requestInfo = {
-      method: httpMethod,
-      headers: {
-        "Content-Type": "application/json",
-        Accept: "application/json",
+    handleRestRequest(
+      partialRestData,
+      restMethods.PATCH,
+      requestUrl,
+      () => {
+        setIsLoaded(false);
+        setIsResolutionComplete(false);
       },
-      body: JSON.stringify(partialRestData),
-    };
-
-    fetch(url, requestInfo)
-      .then((response) => {
-        if (!response.ok) throw new Error(response.status);
-        else return response.json();
-      })
-      .then(
-        (result) => {
-          if (!unmounted.current) {
-            result.oscalSource = JSON.stringify(result, null, "\t");
-            setOscalData(result);
-            setIsLoaded(true);
-          }
-        },
-        (err) => handleFetchError(err)
-      );
+      (result) => {
+        if (!unmounted.current) {
+          result.oscalSource = JSON.stringify(result, null, "\t");
+          setOscalData(result);
+          setIsLoaded(true);
+        }
+      },
+      (err) => handleFetchError(err)
+    );
   };
 
   const handleUrlChange = (event) => {
@@ -227,12 +161,21 @@ export default function OSCALLoader(props) {
 
   const handleRestPut = (jsonString) => {
     handleRestRequest(
-      false,
       JSON.parse(jsonString),
-      null,
-      null,
-      "PUT",
-      oscalUrl
+      restMethods.PUT,
+      oscalUrl,
+      () => {
+        setIsLoaded(false);
+        setIsResolutionComplete(false);
+      },
+      (result) => {
+        if (!unmounted.current) {
+          result.oscalSource = JSON.stringify(result, null, "\t");
+          setOscalData(result);
+          setIsLoaded(true);
+        }
+      },
+      (err) => handleFetchError(err)
     );
   };
 
@@ -366,24 +309,20 @@ export function OSCALCatalogLoader(props) {
       onResolutionComplete={onResolutionComplete}
       onFieldSave={(
         appendToLastFieldInPath,
-        data,
-        editedField,
+        partialRestData,
+        editedFieldJsonPath,
         newValue,
-        restMethod,
         restUrlPath
       ) => {
-        handleRestRequest(
+        handleFieldSave(
           appendToLastFieldInPath,
-          data,
-          editedField,
+          partialRestData,
+          editedFieldJsonPath,
           newValue,
-          restMethod,
           restUrlPath,
-          oscalObjectType.jsonRootName,
-          oscalObjectType.restPath
+          oscalObjectType
         );
       }}
-      restPath={oscalObjectType.restPath}
     />
   );
   return (
@@ -415,24 +354,20 @@ export function OSCALSSPLoader(props) {
       onResolutionComplete={onResolutionComplete}
       onFieldSave={(
         appendToLastFieldInPath,
-        data,
-        editedField,
+        partialRestData,
+        editedFieldJsonPath,
         newValue,
-        restMethod,
         restUrlPath
       ) => {
-        handleRestRequest(
+        handleFieldSave(
           appendToLastFieldInPath,
-          data,
-          editedField,
+          partialRestData,
+          editedFieldJsonPath,
           newValue,
-          restMethod,
           restUrlPath,
-          oscalObjectType.jsonRootName,
-          oscalObjectType.restPath
+          oscalObjectType
         );
       }}
-      restPath={oscalObjectType.restPath}
     />
   );
   return (
@@ -463,24 +398,20 @@ export function OSCALComponentLoader(props) {
       onResolutionComplete={onResolutionComplete}
       onFieldSave={(
         appendToLastFieldInPath,
-        data,
-        editedField,
+        partialRestData,
+        editedFieldJsonPath,
         newValue,
-        restMethod,
         restUrlPath
       ) => {
-        handleRestRequest(
+        handleFieldSave(
           appendToLastFieldInPath,
-          data,
-          editedField,
+          partialRestData,
+          editedFieldJsonPath,
           newValue,
-          restMethod,
           restUrlPath,
-          oscalObjectType.jsonRootName,
-          oscalObjectType.restPath
+          oscalObjectType
         );
       }}
-      restPath={oscalObjectType.restPath}
     />
   );
   return (
@@ -511,24 +442,20 @@ export function OSCALProfileLoader(props) {
       onResolutionComplete={onResolutionComplete}
       onFieldSave={(
         appendToLastFieldInPath,
-        data,
-        editedField,
+        partialRestData,
+        editedFieldJsonPath,
         newValue,
-        restMethod,
         restUrlPath
       ) => {
-        handleRestRequest(
+        handleFieldSave(
           appendToLastFieldInPath,
-          data,
-          editedField,
+          partialRestData,
+          editedFieldJsonPath,
           newValue,
-          restMethod,
           restUrlPath,
-          oscalObjectType.jsonRootName,
-          oscalObjectType.restPath
+          oscalObjectType
         );
       }}
-      restPath={oscalObjectType.restPath}
     />
   );
   return (
